requirements.txt
accelerate==0.26.1
aiofiles==23.2.1
albumentations==1.4.4
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
bitsandbytes==0.43.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
cmake==3.29.0.1
coloredlogs==15.0.1
config==0.5.1
contourpy==1.2.1
cycler==0.12.1
diffusers==0.24.0
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.3
flatbuffers==24.3.25
focal-frequency-loss==0.3.0
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.11.2
lit==18.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.16.2
orjson==3.10.0
packaging==24.0
pandas==2.2.1
Pillow==9.4.0
pip==23.3.1
protobuf==4.25.3
psutil==5.9.8
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-fid==0.3.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.5
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.4.2
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==1.44.1
setproctitle==1.3.3
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torch-fidelity==0.3.0
torchaudio==2.0.2
torchmetrics==1.3.2
torchvision==0.15.2
tqdm==4.64.1
transformers==4.36.2
triton==2.0.0
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing_extensions==4.10.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wandb==0.16.6
websockets==11.0.3
wheel==0.41.2
xformers==0.0.20
zipp==3.18.1
